factor out reactions to messages into react_to()
This prepares the additons of further commands
for which we need more flow control mechanims
than what's possible with inlined conditionals.

Also collect the tokens we recognize for bugzilla/
gerrit ids into {BUGZILLA,GERRIT}_TOKENS. And
also add "bugzilla" to the bugzilla tokens :)

Also carve out the bugzilla/gerrit parsing regexps
into methods. And also take care to escape user
provided parameters when we embed them to regexps.
(Due to this embedding, the parsing regexps can't be
constants -- they need to be presented as instance
variables or methods.) And also refine the regexps,
so that we recognize eg. "bz:34234" (not just
"bz 34234" or "bz: 34234").

Also fix a typo in the Gerrit regexp (change-ids
were allowed to have any letter ([a-z]) but they
can in fact just have a to f ([a-f]).