make gerrit fetching more robust
So far the code assued that answer is either
empty or complete (in the sesne of having the
attributes of a valid change info). However,
some invalid queries result in an answer that
is neither empty nor complete. These were
treated as complete and that broke the code.

Make sure to null out non-complete responses.