cut out responding to messages directed at us
For conveniece, we add a dedicated handler
to messages when addressing the bot as:

  <nick>: <cmd> [<arg>]

So far, in ChanMember, messages were distpacted
to #react_to. Now the primary dispatch target is

  #respond_to(nick, cmd, arg)

if the message is of the format above,
and

  #react_to(nick, msg)

otherwise. #respond_to can instruct to fall
back to #react_to if it returns a string,
which will then be used as the msg
argument of #react_to, or if it returns
:pass, in which case #react_to is called with
the original message.

Bots who wouldn't like to adhere to this
convenience dispatch can just define
their own #read_msg (instead of #react_to).