OO refactor #4: generalize
Certain parts of the BugzillaGerritBot code,
while not part of core bot functionality,
are generic in nature. Now These were extracted
and presented as modules within SimpleIrcBot.

These are: command handling, distinguishing
admin functionality, and caching.

BugzillaGerritBot is now really stripped
down to interacting with these services
and interpreting their output.

Apart from refactor in the sense of moving
code around, the main impact is that command
handling had to be made composable among the
components.

Other semantic change in caching:
take care to distingish false from nil,
so that negative entries can be cached
with false as value.

The new class structure is:

  class SimpleIrcBot
    module Commands
    module Admin
      include Commands
    module Cache
      include Commands
      include Admin
  class BugzillaGerritBot < SimpleIrcBot
    include SimpleIrcBot::Commands
    include SimpleIrcBot::Admin
    include SimpleIrcBot::Cache