add FileOp plugin
... which provides a generic implementation of
loading and savinging a data Hash from/to a
file.

Cache loading/saving was reimplemented with
FileOp. Loading/saving of options was added
with help of FileOp.

Config file argument is now also managed
internally through FileOp -- added
keep_config_file flag to SimpleOptParse
to make this possible.

Added --data-dir option, now any kind of
loading/saving is possible only if it's set,
and the files loaded/saved will be taken from
here. If config/cache file is given in command
line, the default location for saving is their
basename in the data dir. Thus loading/saving
can be enabled both in a way that overwriting
the original files is possible and in a way
it's not.

New class hierarchy:

  class SimpleIrcBot
    module Plugins
      module Commands
      module Admin
        include Commands
      module FileOp
        include Commands
        include Admin
      module Options
        include Commands
        include Admin
        include FileOp
      module Cache
        include Commands
        include Admin
        include FileOp
        include Options
  class BugzillaGerritBot < SimpleIrcBot
    include SimpleIrcBot::Plugins::Commands
    include SimpleIrcBot::Plugins::Admin
    include SimpleIrcBot::Plugins::FileOp
    include SimpleIrcBot::Plugins::Options
    include SimpleIrcBot::Plugins::Cache