be more permissive about character before the service token
So far we demanded, eg. that if a bug is given as
"bz 1111", then before "bz" there should be nothing
or space. Now we allow non-alphanumeric and non-underscore
as well.

This fixes the issue that prefetch-gen output -- which
is a comma-separated list of entities (without space)
was not literally a valid query for the entities listed,