access Bugzilla via API using the ruby-bugzilla gem
The bugzilla lib supports Bugzilla authentication,
so we introduce the ---bugzilla-{user,pass} options.

To honor secrecy of :bugzilla_pass, we add a Hidden constraint flag
in the Options plugin and teach the 'options' command to mask
ouit hidden options. (And of course we add :bugzilla_pass as
hidden option.)