url handling refactor / enhancement
Get rid of mangling user provided values for urls
(that of Bugzilla and Gerrit services).

Rather turn them into URI objects on demand
and get various rendering / compositions
via URI component methods and URI#join.

This also allows us to ket the urls sport
a path component. This can be made use of
when Bugzilla is accessed via API
and that's avialable at a non-standar location.
(We add support for Bugzilla API access in a
subsequent commit.)