fix bug of universal hushing
When a hush window is set, then mention of a given
item makes it hushed for the window in all channels.
However, it should be hushed only on the particular
channel where it was mentioned. We are adding a
chan parameter to the access log that let's us
hush in the contex of that chan.